# Conflicts:
#	Assets/02.Scripts/Drop/Item.cs

+++ Assets/02.Scripts/Rune/Effects/FlameRuneEffect.cs
@@ -9,7 +9,7 @@ public class FlameRuneEffect : ARuneEffect
     public override void Initialize(RuneData data, int tier)
     {
         _tid = data.TID;
-        _damageMultiplier = (int)data.TierList[tier - 1];
+        _damageMultiplier = data.TierList[tier - 1];
     }
 
     public override void ApplyEffect(RuneExecuteContext context, ref Damage damage)
